+++ Events/RingHotspotsEvent.cs
@@ -12,10 +12,10 @@ namespace EddiEvents
         public const string DESCRIPTION = "Triggered when hotspots are detected in a ring";
         public const string SAMPLE = @"{ ""timestamp"":""2019-08-19T00:24:53Z"", ""event"":""SAASignalsFound"", ""BodyName"":""Oponner 6 A Ring"", ""SystemAddress"":3721345878371, ""BodyID"":29, ""Signals"":[ { ""Type"":""Bromellite"", ""Count"":3 }, { ""Type"":""Grandidierite"", ""Count"":5 }, { ""Type"":""LowTemperatureDiamond"", ""Type_Localised"":""Low Temperature Diamonds"", ""Count"":1 } ] }";
 
-        [PublicAPI("The body where surface signals were detected")]
+        [PublicAPI("The ring where hotspots were detected")]
         public string bodyname { get; private set; }
 
-        [PublicAPI("A list of hotspots (as objects)")]
+        [PublicAPI("A list of ring hotspots (as objects with properties 'commodity' and 'amount')")]
         public List<CommodityAmount> hotspots { get; private set; }
 
         // Not intended to be user facing